Fixes an issue with late close of file handles.
Although in the command running loop we defer'd the close of stdout, so
we technically were safe, because that was in a loop the defer wasn't
closed until a lot later than necessary, and on large batches can cause
ulimit to be exceeded.

This fix moves the command exection to its own function which fixes the
late defer, and also makes the code more easy to read.